ui: remove antipatterns from EventStartDateFilter * ref: https://github.com/cern-sis/issues-inspire/issues/1707
